Another compile fix for Win64.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@22099 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/software/sft3dcom.cpp
@@ -4364,7 +4364,7 @@ void csSoftwareGraphics3DCommon::DrawSimpleMesh (const csSimpleRenderMesh &mesh,
   {
     csRenderBufferLock<uint> indexLock (scrapIndices);
     for (uint i = 0; i < mesh.vertexCount; i++)
-      indexLock[i] = i;
+      indexLock[(size_t)i] = i;
   }
   sv->SetValue (scrapIndices);
   scrapBufferHolder->SetRenderBuffer (CS_BUFFER_INDEX, scrapIndices);
